+++ mithril-aggregator/src/services/message.rs
@@ -6,8 +6,9 @@ use async_trait::async_trait;
 use thiserror::Error;
 
 use mithril_common::{
-    entities::SignedEntityTypeDiscriminants,
+    entities::{Epoch, SignedEntityTypeDiscriminants},
     messages::{
+        CardanoStakeDistributionListMessage, CardanoStakeDistributionMessage,
         CardanoTransactionSnapshotListMessage, CardanoTransactionSnapshotMessage,
         CertificateListMessage, CertificateMessage, MithrilStakeDistributionListMessage,
         MithrilStakeDistributionMessage, SnapshotListMessage, SnapshotMessage,
@@ -74,6 +75,24 @@ pub trait MessageService: Sync + Send {
         &self,
         limit: usize,
     ) -> StdResult<CardanoTransactionSnapshotListMessage>;
+
+    /// Return the information regarding the Cardano stake distribution for the given identifier.
+    async fn get_cardano_stake_distribution_message(
+        &self,
+        signed_entity_id: &str,
+    ) -> StdResult<Option<CardanoStakeDistributionMessage>>;
+
+    /// Return the information regarding the Cardano stake distribution for the given epoch.
+    async fn get_cardano_stake_distribution_message_by_epoch(
+        &self,
+        epoch: Epoch,
+    ) -> StdResult<Option<CardanoStakeDistributionMessage>>;
+
+    /// Return the list of the last Cardano stake distributions message
+    async fn get_cardano_stake_distribution_list_message(
+        &self,
+        limit: usize,
+    ) -> StdResult<CardanoStakeDistributionListMessage>;
 }
 
 /// Implementation of the [MessageService]
@@ -186,6 +205,43 @@ impl MessageService for MithrilMessageService {
 
         entities.into_iter().map(|i| i.try_into()).collect()
     }
+
+    async fn get_cardano_stake_distribution_message(
+        &self,
+        signed_entity_id: &str,
+    ) -> StdResult<Option<CardanoStakeDistributionMessage>> {
+        let signed_entity = self
+            .signed_entity_storer
+            .get_signed_entity(signed_entity_id)
+            .await?;
+
+        signed_entity.map(|v| v.try_into()).transpose()
+    }
+
+    async fn get_cardano_stake_distribution_message_by_epoch(
+        &self,
+        epoch: Epoch,
+    ) -> StdResult<Option<CardanoStakeDistributionMessage>> {
+        let signed_entity = self
+            .signed_entity_storer
+            .get_cardano_stake_distribution_signed_entity_by_epoch(epoch)
+            .await?;
+
+        signed_entity.map(|v| v.try_into()).transpose()
+    }
+
+    async fn get_cardano_stake_distribution_list_message(
+        &self,
+        limit: usize,
+    ) -> StdResult<CardanoStakeDistributionListMessage> {
+        let signed_entity_type_id = SignedEntityTypeDiscriminants::CardanoStakeDistribution;
+        let entities = self
+            .signed_entity_storer
+            .get_last_signed_entities_by_type(&signed_entity_type_id, limit)
+            .await?;
+
+        entities.into_iter().map(|i| i.try_into()).collect()
+    }
 }
 
 #[cfg(test)]
@@ -193,8 +249,8 @@ mod tests {
     use std::sync::Arc;
 
     use mithril_common::entities::{
-        CardanoTransactionsSnapshot, Certificate, Epoch, MithrilStakeDistribution, SignedEntity,
-        SignedEntityType, Snapshot,
+        CardanoStakeDistribution, CardanoTransactionsSnapshot, Certificate, Epoch,
+        MithrilStakeDistribution, SignedEntity, SignedEntityType, Snapshot,
     };
     use mithril_common::messages::ToMessageAdapter;
     use mithril_common::test_utils::MithrilFixtureBuilder;
@@ -203,6 +259,7 @@ mod tests {
     use crate::database::repository::MockSignedEntityStorer;
     use crate::dependency_injection::DependenciesBuilder;
     use crate::message_adapters::{
+        ToCardanoStakeDistributionListMessageAdapter, ToCardanoStakeDistributionMessageAdapter,
         ToCardanoTransactionListMessageAdapter, ToCardanoTransactionMessageAdapter,
         ToMithrilStakeDistributionListMessageAdapter, ToMithrilStakeDistributionMessageAdapter,
         ToSnapshotListMessageAdapter, ToSnapshotMessageAdapter,
@@ -494,6 +551,130 @@ mod tests {
         let service = dep_builder.get_message_service().await.unwrap();
         let response = service
             .get_cardano_transaction_list_message(10)
+            .await
+            .unwrap();
+
+        assert_eq!(message, response);
+    }
+
+    #[tokio::test]
+    async fn get_cardano_stake_distribution() {
+        let entity = SignedEntity::<CardanoStakeDistribution>::dummy();
+        let record = SignedEntityRecord {
+            signed_entity_id: entity.signed_entity_id.clone(),
+            signed_entity_type: SignedEntityType::CardanoStakeDistribution(entity.artifact.epoch),
+            certificate_id: entity.certificate_id.clone(),
+            artifact: serde_json::to_string(&entity.artifact).unwrap(),
+            created_at: entity.created_at,
+        };
+        let message = ToCardanoStakeDistributionMessageAdapter::adapt(entity);
+        let configuration = Configuration::new_sample();
+        let mut dep_builder = DependenciesBuilder::new(configuration);
+        let mut storer = MockSignedEntityStorer::new();
+        storer
+            .expect_get_signed_entity()
+            .return_once(|_| Ok(Some(record)))
+            .once();
+        dep_builder.signed_entity_storer = Some(Arc::new(storer));
+        let service = dep_builder.get_message_service().await.unwrap();
+        let response = service
+            .get_cardano_stake_distribution_message("whatever")
+            .await
+            .unwrap()
+            .expect("A CardanoStakeDistributionMessage was expected.");
+
+        assert_eq!(message, response);
+    }
+
+    #[tokio::test]
+    async fn get_cardano_stake_distribution_not_exist() {
+        let configuration = Configuration::new_sample();
+        let mut dep_builder = DependenciesBuilder::new(configuration);
+        let mut storer = MockSignedEntityStorer::new();
+        storer
+            .expect_get_signed_entity()
+            .return_once(|_| Ok(None))
+            .once();
+        dep_builder.signed_entity_storer = Some(Arc::new(storer));
+        let service = dep_builder.get_message_service().await.unwrap();
+        let response = service
+            .get_cardano_stake_distribution_message("whatever")
+            .await
+            .unwrap();
+
+        assert!(response.is_none());
+    }
+
+    #[tokio::test]
+    async fn get_cardano_stake_distribution_by_epoch() {
+        let entity = SignedEntity::<CardanoStakeDistribution>::dummy();
+        let record = SignedEntityRecord {
+            signed_entity_id: entity.signed_entity_id.clone(),
+            signed_entity_type: SignedEntityType::CardanoStakeDistribution(entity.artifact.epoch),
+            certificate_id: entity.certificate_id.clone(),
+            artifact: serde_json::to_string(&entity.artifact).unwrap(),
+            created_at: entity.created_at,
+        };
+        let message = ToCardanoStakeDistributionMessageAdapter::adapt(entity.clone());
+        let configuration = Configuration::new_sample();
+        let mut dep_builder = DependenciesBuilder::new(configuration);
+        let mut storer = MockSignedEntityStorer::new();
+        storer
+            .expect_get_cardano_stake_distribution_signed_entity_by_epoch()
+            .return_once(|_| Ok(Some(record)))
+            .once();
+        dep_builder.signed_entity_storer = Some(Arc::new(storer));
+        let service = dep_builder.get_message_service().await.unwrap();
+        let response = service
+            .get_cardano_stake_distribution_message_by_epoch(Epoch(999))
+            .await
+            .unwrap()
+            .expect("A CardanoStakeDistributionMessage was expected.");
+
+        assert_eq!(message, response);
+    }
+
+    #[tokio::test]
+    async fn get_cardano_stake_distribution_by_epoch_not_exist() {
+        let configuration = Configuration::new_sample();
+        let mut dep_builder = DependenciesBuilder::new(configuration);
+        let mut storer = MockSignedEntityStorer::new();
+        storer
+            .expect_get_cardano_stake_distribution_signed_entity_by_epoch()
+            .return_once(|_| Ok(None))
+            .once();
+        dep_builder.signed_entity_storer = Some(Arc::new(storer));
+        let service = dep_builder.get_message_service().await.unwrap();
+        let response = service
+            .get_cardano_stake_distribution_message_by_epoch(Epoch(999))
+            .await
+            .unwrap();
+
+        assert!(response.is_none());
+    }
+
+    #[tokio::test]
+    async fn get_cardano_stake_distribution_list_message() {
+        let entity = SignedEntity::<CardanoStakeDistribution>::dummy();
+        let records = vec![SignedEntityRecord {
+            signed_entity_id: entity.signed_entity_id.clone(),
+            signed_entity_type: SignedEntityType::CardanoStakeDistribution(entity.artifact.epoch),
+            certificate_id: entity.certificate_id.clone(),
+            artifact: serde_json::to_string(&entity.artifact).unwrap(),
+            created_at: entity.created_at,
+        }];
+        let message = ToCardanoStakeDistributionListMessageAdapter::adapt(vec![entity]);
+        let configuration = Configuration::new_sample();
+        let mut dep_builder = DependenciesBuilder::new(configuration);
+        let mut storer = MockSignedEntityStorer::new();
+        storer
+            .expect_get_last_signed_entities_by_type()
+            .return_once(|_, _| Ok(records))
+            .once();
+        dep_builder.signed_entity_storer = Some(Arc::new(storer));
+        let service = dep_builder.get_message_service().await.unwrap();
+        let response = service
+            .get_cardano_stake_distribution_list_message(10)
             .await
             .unwrap();