+++ mithril-common/src/test_utils/fake_data.rs
@@ -7,7 +7,7 @@ use crate::crypto_helper::{self, ProtocolMultiSignature};
 use crate::entities::{
     self, BlockNumber, CertificateMetadata, CertificateSignature, CompressionAlgorithm, Epoch,
     LotteryIndex, ProtocolMessage, ProtocolMessagePartKey, SignedEntityType, SingleSignatures,
-    SlotNumber, StakeDistributionParty,
+    SlotNumber, StakeDistribution, StakeDistributionParty,
 };
 use crate::test_utils::MithrilFixtureBuilder;
 
@@ -202,7 +202,8 @@ pub fn snapshots(total: u64) -> Vec<entities::Snapshot> {
     (1..total + 1)
         .map(|snapshot_id| {
             let digest = format!("1{snapshot_id}").repeat(20);
-            let beacon = beacon();
+            let mut beacon = beacon();
+            beacon.immutable_file_number += snapshot_id;
             let certificate_hash = "123".to_string();
             let size = snapshot_id * 100000;
             let cardano_node_version = Version::parse("1.0.0").unwrap();
@@ -257,4 +258,21 @@ pub const fn transaction_hashes<'a>() -> [&'a str; 5] {
         "3f6f3c981c89097f62c9b43632875db7a52183ad3061c822d98259d18cd63dcf",
         "f4fd91dccc25fd63f2caebab3d3452bc4b2944fcc11652214a3e8f1d32b09713",
     ]
+}
+
+/// Fake Cardano Stake Distributions
+pub fn cardano_stake_distributions(total: u64) -> Vec<entities::CardanoStakeDistribution> {
+    (1..total + 1)
+        .map(|epoch_idx| cardano_stake_distribution(Epoch(epoch_idx)))
+        .collect::<Vec<entities::CardanoStakeDistribution>>()
+}
+
+/// Fake Cardano Stake Distribution
+pub fn cardano_stake_distribution(epoch: Epoch) -> entities::CardanoStakeDistribution {
+    let stake_distribution = StakeDistribution::from([("pool-1".to_string(), 100)]);
+    entities::CardanoStakeDistribution {
+        hash: format!("hash-epoch-{epoch}"),
+        epoch,
+        stake_distribution,
+    }
 }